game-screen refactor draft
